Add to-matcher and rewriting of "From" header
These are really one use-case, hence the double-feature in one commit:
PHP's shitty form mailers tend to fake the "From" header, which won't
work with SMTP servers that are actually careful about who is allowed to
send an email as a given user.  So when "To" matches the form mailer's
"to" address, we can use an email address like "contact@..." as the
"from" address.