email: undo crazy header restrictions
Headers are now mutable and easily accessible.  The aliased magic that
made this a huge pain is now gone.  Instead of holding "from", "to",
etc. in local fields, they just have to be accessed from the email's
header, which means more code for callers, but a *much* simpler setup
for the email package since we're not trying to figure out how to keep
the header and those values in sync.